+++ lib/middleware.dart
@@ -1,0 +1,5 @@
+/// Additional utilities.
+library;
+
+export 'src/middleware.dart';
+export 'src/tag_parsing.dart';