+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial version.
+
+### 2.0.0
+- Added and deprecated credential members
+- Added UI colors
+- Added and removed tag parsing members